% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bireg.R
\name{bca}
\alias{bca}
\title{Bi-clustering analysis to predicted the regulated patients and genes}
\usage{
bca(mx, regulators, min.pas = 100, min.ges = 50, cutoff = 0.85,
  cores = 1, max.ges = 1000)
}
\arguments{
\item{mx}{an expression matrix}

\item{regulators}{a vector for the index or gene name of regulators}

\item{min.pas}{the minimum number of regulated patients}

\item{min.ges}{the minimum number of regulated genes}

\item{cutoff}{the cutoff value of Spearman's correlation}

\item{cores}{the number of CPU cores}

\item{max.ges}{the maximum number of regulated genes}
}
\value{
A bca class object.
}
\description{
Bi-clustering analysis to predicted the regulated patients and genes
}
\details{
Use a cpp program to do bi-clustering analysis to predict the genes and patients regulated by regulators
}
\examples{
res <- bca(expr, regulators, min.pas=50, min.ges=10, max.ges=60)

}
\references{

}
\author{
Guofeng Meng
}
