% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bireg.R
\name{extract.modules}
\alias{extract.modules}
\title{Extract regulated patients and genes}
\usage{
extract.modules(bca.obj, expr, mod = "max.gene")
}
\arguments{
\item{bca.obj}{a ``bca'' object generated by "bca"}

\item{expr}{an expression matrix}

\item{mod}{the mode for extract the regulated patients and genes. It can be "max.genes" or "max.patients"}
}
\value{
a list
}
\description{
Extract regulated patients and genes
}
\details{
Use a cpp program to do bi-clustering analysis to predict the genes and patients regulated by regulators
}
\examples{
\dontrun{
output <- extract.modules(res, expr, "max.genes")
}
}
\references{

}
\author{
Guofeng Meng
}
